fixed some server bugs

+++ controllers/feed.js
@@ -69,8 +69,8 @@ exports.createPost = (req, res, next) => {
     })
     .then(result => {
       res.status(201).json({
-        message: 'Post created succesfully!',
-        posts: post,
+        message: 'Post created successfully!',
+        post: post,
         creator: { _id: creator._id, name: creator.name },
       });
     })
@@ -165,12 +165,25 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized');
+        error.statusCode = 403;
+        throw error;
+      }
+
       //Check logged in user
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
     .then(result => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(result => {
       res.status(200).json({ message: 'Deleted post.' });
     })
     .catch(err => {
